% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adjust_outliers.R
\name{adjust_outliers}
\alias{adjust_outliers}
\title{Adjusts outliers based off of geom_boxplot}
\usage{
adjust_outliers(dataframe, column)
}
\arguments{
\item{dataframe}{Object of data table or dataframe}

\item{column}{Column in questions with double quotes}
}
\description{
This function identifies and plots outliers
}
\examples{
Example <- get_outlier_plot()
}
