% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adjust_outliers_knn.R
\name{adjust_outliers_knn}
\alias{adjust_outliers_knn}
\title{Adjusts outliers based off of geom_boxplot}
\usage{
adjust_outliers_knn(dataframe, column)
}
\arguments{
\item{dataframe}{Object of data table or dataframe}

\item{column}{Column in question with double quotes}
}
\description{
This function adjusts outliers using the kNN algorithm
}
