% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_outlier_plot.R
\name{get_outlier_plot}
\alias{get_outlier_plot}
\title{Identifies and adjust outliers in data}
\usage{
get_outlier_plot(dataframe, column)
}
\arguments{
\item{dataframe}{Object of data table or dataframe}

\item{column}{Column in questions with double quotes}
}
\description{
This function identifies and plots outliers
}
\examples{
Example <- get_outlier_plot()
}
