% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adjust_outliers.R
\name{adjust_outliers_knn}
\alias{adjust_outliers_knn}
\title{Adjusts outliers based off of geom_boxplot}
\usage{
adjust_outliers_knn(dataframe, column)
}
\arguments{
\item{dataframe}{Object of data table or dataframe}

\item{column}{Column in question with double quotes}
}
\description{
This function adjusts outliers using the kNN algorithm
}
\examples{
Example <- adjust_outliers()

The original and imputed series can be compared with the below

plot<- Final_Data \%>\% 
ggplot(aes(x=Date,y=Imputed)) + #Where Imputed is the name of the imputed series
geom_line(col="black") +
geom_point()

comp_plot<- plot + geom_line(aes(x=Date,y=Original), col = "red") + 
geom_point() + ggtitle("Original vs Imputed Series") + ylab("Series") #Where Original is the name of the Original series
                  
ggplotly(comp_plot)

}
